// themes/cerveza/variables.d.ts
export declare const color: {
    brand: {
        primary: string;
        secondary: string;
        tertiary: string;
    };
    blueGray: {
        darkest: string;
        darker: string;
        dark: string;
        main: string;
        light: string;
        lighter: string;
        lightest: string;
    };
    gray: {
        darkest: string;
        darker: string;
        dark: string;
        main: string;
        light: string;
        lighter: string;
        lightest: string;
    };
    blue: {
        dark: string;
        main: string;
        light: string;
        lighter: string;
        lightest: string;
    };
    orange: {
        dark: string;
        main: string;
        light: string;
        lighter: string;
    };
    accent: {
        success: string;
        info: string;
        warning: string;
        error: string;
        shared: string;
    };
};
export declare const size: {
    xxsmall: string;
    xsmall: string;
    small: string;
    normal: string;
    large: string;
    xlarge: string;
    xxlarge: string;
};